# docs/Worth_It_OpenAPI_v1.0.yaml
openapi: 3.1.0
info:
  title: Worth It API
  version: 1.0.0
  description: Internal REST API contract for the Worth It mobile app MVP. All monetary amounts use integer minor
    units plus an ISO 4217 currency code.
servers:
- url: https://api.worthit.app/v1
  description: Production
- url: https://staging-api.worthit.app/v1
  description: Staging
security:
- bearerAuth: []
...
paths:
...
  /scans:
    post:
      summary: Create scan
      tags:
      - Scans
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scan'
        '402':
          description: No scan entitlement
        '409':
          description: Idempotency conflict
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScanInput'
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
    get:
      summary: List saved scans
      tags:
      - Scans
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Scan'
                  next_cursor:
                    type: string
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /scans/{scan_id}:
    get:
      summary: Get scan
      tags:
      - Scans
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scan'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: scan_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      summary: Delete scan
      tags:
      - Scans
      responses:
        '204':
          description: Deleted
        '401':
          description: Unauthorized
        '404':
          description: Not found
      parameters:
      - name: scan_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/IdempotencyKey'
  /scans/{scan_id}/upload-urls:
    post:
      summary: Create signed image upload URLs
      tags:
      - Scans
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  uploads:
                    type: array
                    items:
                      type: object
                      properties:
                        image_id:
                          type: string
                          format: uuid
                        upload_url:
                          type: string
                          format: uri
                        object_path:
                          type: string
                        expires_at:
                          type: string
                          format: date-time
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - files
              properties:
                files:
                  type: array
                  minItems: 1
                  maxItems: 6
                  items:
                    type: object
                    required:
                    - content_type
                    - size_bytes
                    properties:
                      content_type:
                        type: string
                        enum:
                        - image/jpeg
                        - image/png
                        - image/heic
                      size_bytes:
                        type: integer
                        maximum: 12582912
      parameters:
      - name: scan_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/IdempotencyKey'
  /scans/{scan_id}/identify:
    post:
      summary: Start item identification
      tags:
      - Analysis
      responses:
        '202':
          description: Identification started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scan'
        '409':
          description: Invalid scan state
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                barcode:
                  type: string
      parameters:
      - name: scan_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/IdempotencyKey'
  /scans/{scan_id}/item:
    patch:
      summary: Confirm or correct item details
      tags:
      - Scans
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scan'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemDetails'
      parameters:
      - name: scan_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/IdempotencyKey'
  /scans/{scan_id}/analyze:
    post:
      summary: Start market analysis
      tags:
      - Analysis
      responses:
        '202':
          description: Analysis accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scan'
        '402':
          description: No scan entitlement
        '409':
          description: Invalid scan state or duplicate request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                target_profit:
                  $ref: '#/components/schemas/Money'
                target_roi_percent:
                  type: number
      parameters:
      - name: scan_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/IdempotencyKey'
  /scans/{scan_id}/cancel:
    post:
      summary: Cancel active analysis
      tags:
      - Analysis
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scan'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: {}
      parameters:
      - name: scan_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/IdempotencyKey'
  /scans/{scan_id}/comparables:
    get:
      summary: List analysis comparables
      tags:
      - Comparables
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Comparable'
                  data_freshness_at:
                    type: string
                    format: date-time
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: scan_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    webhookSignature:
      type: apiKey
      in: header
      name: X-Webhook-Signature
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
...
  schemas:
...
    Error:
      type: object
      required:
      - error
      - request_id
      properties:
        error:
          type: object
          required:
          - code
          - message
          - retryable
          properties:
            code:
              type: string
            message:
              type: string
            retryable:
              type: boolean
            details:
              type: object
              additionalProperties: true
        request_id:
          type: string
        timestamp:
          type: string
          format: date-time
...
    ScanInput:
      type: object
      required:
      - preferred_currency
      properties:
        preferred_currency:
          type: string
          pattern: ^[A-Z]{3}$
        purchase_price:
          $ref: '#/components/schemas/Money'
        category_hint:
          type: string
        source_context:
          type: string
          enum:
          - thrift_store
          - marketplace
          - garage_sale
          - swap_meet
          - other
    ItemDetails:
      type: object
      required:
      - title
      - category
      - condition
      properties:
        title:
          type: string
        brand:
          type: string
        model:
          type: string
        category:
          type: string
        condition:
          type: string
          enum:
          - new
          - open_box
          - excellent
          - good
          - fair
          - parts_or_repair
        attributes:
          type: object
          additionalProperties:
            type: string
        accessories:
          type: array
          items:
            type: string
        repair_cost:
          $ref: '#/components/schemas/Money'
        outbound_shipping:
          $ref: '#/components/schemas/Money'
        taxes_paid:
          $ref: '#/components/schemas/Money'
        notes:
          type: string
...
    Scan:
      type: object
      required:
      - id
      - status
      - created_at
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum:
          - draft
          - uploaded
          - identifying
          - ready
          - analyzing
          - completed
          - identification_failed
          - analysis_failed
          - cancelled
        input:
          $ref: '#/components/schemas/ScanInput'
        item:
          $ref: '#/components/schemas/ItemDetails'
        analysis:
          $ref: '#/components/schemas/AnalysisResult'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...